wip: initial header detection
